Code refactoring
ResultServlet was refactored (MainPageServlet has been renamed to ResultServlet). Improved data presentation in result.jsp (now there is a scroll bar there).
ApproximationResultDto was added and that is used for the data transport between Serlvets and Services.
OutputConverter was added. It is used to convert Double list (such as argumentValues) to String
list according the OUTPUT_FORMAT. The String lists is used by result.jsp.