Filter has been added
ResultFilter is used to validate user's data for the approximation. If
some data is incorrect a message appears and shows that some fields
contain invalid data.
Validator was created for the data validation.
ApproximationParametersValidator implements Validator and itself is used
by ResultFilter.
A button in result.jsp was added. It is used to return to the inital
page (index.jsp).