Code refactoring and a few changes
ApproximationParameter, ApproximationParametersValidator were renamed.
Some fields and variables also were renamed.
index.jsp was put in the WEB-INF directory. IndexServlet was created to access the index.jsp. IndexFilter was created as well.
PathHelper was created and that is used to get the path to the .jsp files using
theirs names.